requirements.txt
firebase-admin==6.0.1
Flask==2.2.2
Flask-Cors==3.0.10
flask-restx==1.0.3
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
